DDCNL-9021 | moved /date-of-marriage into its own controller
