+++ app/views/date_of_marriage.scala.html
@@ -47,7 +47,7 @@ request: AuthenticatedUserRequest[_])
   @p(HtmlContent(messages("pages.date-of-marriage.para1")))
   @p(HtmlContent(messages("pages.date-of-marriage.para2")))
 
-  @form(action = routes.TransferController.dateOfMarriageAction(), args = (Symbol("id") -> "date-of-marriage-form")) {
+  @form(action = transfer.routes.DateOfMarriageController.dateOfMarriageAction(), args = (Symbol("id") -> "date-of-marriage-form")) {
     @inputDate(
         fieldName = "dateOfMarriage",
         form = marriageForm,
