+++ conf/app.routes
@@ -16,10 +16,10 @@ POST        /benefit-calculator-pta             controllers.EligibilityCalculato
 # Transfer Allowance
 GET         /transfer-allowance                 controllers.transfer.TransferAllowanceController.transfer()
 POST        /transfer-allowance                 controllers.transfer.TransferAllowanceController.transferAction()
-GET         /date-of-marriage                   controllers.TransferController.dateOfMarriage()
-GET         /date-of-marriage-en                controllers.TransferController.dateOfMarriageWithEn()
-GET         /date-of-marriage-cy                controllers.TransferController.dateOfMarriageWithCy()
-POST        /date-of-marriage                   controllers.TransferController.dateOfMarriageAction()
+GET         /date-of-marriage                   controllers.transfer.DateOfMarriageController.dateOfMarriage()
+GET         /date-of-marriage-en                controllers.transfer.DateOfMarriageController.dateOfMarriageWithEn()
+GET         /date-of-marriage-cy                controllers.transfer.DateOfMarriageController.dateOfMarriageWithCy()
+POST        /date-of-marriage                   controllers.transfer.DateOfMarriageController.dateOfMarriageAction()
 GET         /eligible-years                     controllers.TransferController.eligibleYears()
 POST        /eligible-years                     controllers.TransferController.eligibleYearsAction()
 GET         /previous-years                     controllers.TransferController.previousYears()
