+++ test/controllers/transfer/DateOfMarriageControllerTest.scala
@@ -30,16 +30,17 @@ import play.api.inject.bind
 import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
+import services.CacheService.CACHE_MARRIAGE_DATE
 import services.{CachingService, TimeService, TransferService}
 import test_utils.TestData.Ninos
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.emailaddress.EmailAddress
 import uk.gov.hmrc.time
 import utils.{ControllerBaseTest, MockAuthenticatedAction}
-import services.CacheService._
+
 import java.time.LocalDate
 
-class TransferAllowanceControllerTest extends ControllerBaseTest {
+class DateOfMarriageControllerTest extends ControllerBaseTest {
 
   val currentTaxYear: Int = time.TaxYear.current.startYear
   val mockTransferService: TransferService = mock[TransferService]
@@ -59,45 +60,74 @@ class TransferAllowanceControllerTest extends ControllerBaseTest {
     )
     .build()
 
-  def controller: TransferAllowanceController =
-    app.injector.instanceOf[TransferAllowanceController]
+  def controller: DateOfMarriageController =
+    app.injector.instanceOf[DateOfMarriageController]
 
   when(mockTimeService.getCurrentDate) thenReturn LocalDate.now()
   when(mockTimeService.getCurrentTaxYear) thenReturn currentTaxYear
 
-  "transfer" should {
+
+  "dateOfMarriage" should {
     "return success" in {
-      val result = controller.transfer()(request)
+      val result = controller.dateOfMarriage()(request)
       status(result) shouldBe OK
     }
   }
 
-  "transferAction" should {
+  "dateOfMarriageWithCy" should {
+    "redirect to dateOfMarriage, with a welsh language setting" in {
+      val result = await(controller.dateOfMarriageWithCy()(request))
+      status(result)               shouldBe SEE_OTHER
+      redirectLocation(result)     shouldBe Some(controllers.transfer.routes.DateOfMarriageController.dateOfMarriage().url)
+      result.newCookies.head.name  shouldBe "PLAY_LANG"
+      result.newCookies.head.value shouldBe "cy"
+    }
+  }
+
+  "dateOfMarriageWithEn" should {
+    "redirect to dateOfMarriage, with an english language setting" in {
+      val result = await(controller.dateOfMarriageWithEn()(request))
+      status(result)               shouldBe SEE_OTHER
+      redirectLocation(result)     shouldBe Some(controllers.transfer.routes.DateOfMarriageController.dateOfMarriage().url)
+      result.newCookies.head.name  shouldBe "PLAY_LANG"
+      result.newCookies.head.value shouldBe "en"
+    }
+  }
+
+  "dateOfMarriageAction" should {
     "return bad request" when {
       "an invalid form is submitted" in {
-        val recipientDetails: RecipientDetailsFormInput =
-          RecipientDetailsFormInput("Test", "User", Gender("M"), Nino(Ninos.nino2))
-        when(mockCachingService.put[RecipientDetailsFormInput](ArgumentMatchers.eq(CACHE_RECIPIENT_DETAILS), ArgumentMatchers.eq(recipientDetails))(any(), any()))
-          .thenReturn(recipientDetails)
-        val result = controller.transferAction()(request)
+        val result = controller.dateOfMarriageAction()(request)
         status(result) shouldBe BAD_REQUEST
       }
     }
+
     "redirect the user" when {
       "a valid form is submitted" in {
-        val recipientDetails: RecipientDetailsFormInput =
-          RecipientDetailsFormInput("Test", "User", Gender("M"), Nino(Ninos.nino2))
+        val dateOfMarriageInput = DateOfMarriageFormInput(LocalDate.now().minusDays(1))
         val request = FakeRequest().withMethod("POST").withFormUrlEncodedBody(
-          "name"      -> "Test",
-          "last-name" -> "User",
-          "gender"    -> "M",
-          "nino"      -> Ninos.nino2
+          "dateOfMarriage.year"  -> dateOfMarriageInput.dateOfMarriage.getYear.toString,
+          "dateOfMarriage.month" -> s"0${dateOfMarriageInput.dateOfMarriage.getMonthValue.toString}",
+          "dateOfMarriage.day"   -> dateOfMarriageInput.dateOfMarriage.getDayOfMonth.toString
         )
-        when(mockCachingService.put[RecipientDetailsFormInput](ArgumentMatchers.eq(CACHE_RECIPIENT_DETAILS), ArgumentMatchers.eq(recipientDetails))(any(), any()))
-          .thenReturn(recipientDetails)
-        val result = controller.transferAction()(request)
+        val registrationFormInput =
+          RegistrationFormInput("Test", "User", Gender("F"), Nino(Ninos.nino1), dateOfMarriageInput.dateOfMarriage)
+
+        when(mockCachingService.put[DateOfMarriageFormInput](ArgumentMatchers.eq(CACHE_MARRIAGE_DATE), ArgumentMatchers.eq(dateOfMarriageInput))(any(), any()))
+          .thenReturn(dateOfMarriageInput)
+
+        when(mockTransferService.getRecipientDetailsFormData()(any(), any(), any()))
+          .thenReturn(RecipientDetailsFormInput("Test", "User", Gender("F"), Nino(Ninos.nino1)))
+        when(
+          mockTransferService.isRecipientEligible(
+            ArgumentMatchers.eq(Nino(Ninos.nino1)),
+            ArgumentMatchers.eq(registrationFormInput)
+          )(any(), any(), any())
+        )
+          .thenReturn(true)
+        val result = controller.dateOfMarriageAction()(request)
         status(result)           shouldBe SEE_OTHER
-        redirectLocation(result) shouldBe Some(controllers.transfer.routes.DateOfMarriageController.dateOfMarriage().url)
+        redirectLocation(result) shouldBe Some(controllers.routes.TransferController.eligibleYears().url)
       }
     }
   }
